requirements.txt
bokeh==3.6.0
contourpy==1.3.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyelftools==0.31
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
ruff==0.7.0
six==1.16.0
squarify==0.4.4
tornado==6.4.1
tzdata==2024.2
xyzservices==2024.9.0
